sales_agent_server/pyproject.toml
[project]
name = "sales-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "authlib>=1.6.0",
    "chromadb>=1.0.15",
    "fastapi>=0.115.14",
    "google-genai>=1.24.0",
    "google-generativeai>=0.8.5",
    "itsdangerous>=2.2.0",
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "langchain-google-genai>=2.0.10",
    "langchain-groq>=0.3.5",
    "langgraph>=0.5.1",
    "psycopg2>=2.9.10",
    "pymupdf>=1.26.3",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.1",
    "python-jose>=3.5.0",
    "uvicorn>=0.35.0",
]
